requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-celery-results==2.5.1
django-filter==24.3
django-redis==5.4.0

django-resized==1.0.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
Markdown==3.7
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
